broker submit invoice to tax module

+++ broker/src/main/java/pt/ulisboa/tecnico/softeng/broker/interfaces/BankInterface.java
@@ -4,7 +4,7 @@ import pt.ulisboa.tecnico.softeng.bank.dataobjects.BankOperationData;
 import pt.ulisboa.tecnico.softeng.bank.domain.Bank;
 
 public class BankInterface {
-	public static String processPayment(String IBAN, int amount) {
+	public static String processPayment(String IBAN, double amount) {
 		return Bank.processPayment(IBAN, amount);
 	}
